requirements.txt
setuptools>=40.6.3
PyYAML>=5.1